+++ Source/OnlineSubsystemAccelByte/Private/OnlineEntitlementsInterfaceAccelByte.cpp
@@ -8,6 +8,7 @@
 #include "AsyncTasks/Entitlements/OnlineAsyncTaskAccelByteConsumeEntitlement.h"
 #include "AsyncTasks/Entitlements/OnlineAsyncTaskAccelByteSyncPlatformPurchase.h"
 #include "AsyncTasks/Entitlements/OnlineAsyncTaskAccelByteSyncDLC.h"
+#include "AsyncTasks/Entitlements/OnlineAsyncTaskAccelByteGetUserEntitlementHistory.h"
 
 FOnlineEntitlementsAccelByte::FOnlineEntitlementsAccelByte(FOnlineSubsystemAccelByte* InSubsystem)
 	: AccelByteSubsystem(InSubsystem)
@@ -22,6 +23,17 @@ void FOnlineEntitlementsAccelByte::AddEntitlementToMap(const TSharedRef<const FU
 
 	EntMap.Emplace(Entitlement->Id, Entitlement);
 	ItemEntMap.Emplace(Entitlement->ItemId, Entitlement);
+}
+
+void FOnlineEntitlementsAccelByte::AddUserEntitlementHistoryToMap(const TSharedRef<const FUniqueNetIdAccelByteUser>& UserId
+	, const FUniqueEntitlementId& EntitlementId
+	, TArray<FAccelByteModelsUserEntitlementHistory> UserEntitlementHistory)
+{
+	FScopeLock ScopeLock(&EntitlementMapLock);
+
+	FEntitlementHistory& EntitlementHistoryMap = UserEntitlementHistoryMap.FindOrAdd(UserId);
+
+	EntitlementHistoryMap.Emplace(EntitlementId, UserEntitlementHistory);
 }
 
 bool FOnlineEntitlementsAccelByte::GetFromSubsystem(const IOnlineSubsystem* Subsystem, FOnlineEntitlementsAccelBytePtr& OutInterfaceInstance)
@@ -104,4 +116,68 @@ void FOnlineEntitlementsAccelByte::SyncDLC(const FUniqueNetId& InLocalUserId, co
 void FOnlineEntitlementsAccelByte::ConsumeEntitlement(const FUniqueNetId& UserId, const FUniqueEntitlementId& EntitlementId, int32 UseCount, TArray<FString> Options, const FString& RequestId)
 {
 	AccelByteSubsystem->CreateAndDispatchAsyncTaskParallel<FOnlineAsyncTaskAccelByteConsumeEntitlement>(AccelByteSubsystem, UserId, EntitlementId, UseCount, Options, RequestId);
+}
+
+void FOnlineEntitlementsAccelByte::GetUserEntitlementHistory(int32 LocalUserNum, const FUniqueEntitlementId& EntitlementId, bool bForceUpdate)
+{
+	AccelByteSubsystem->CreateAndDispatchAsyncTaskParallel<FOnlineAsyncTaskAccelByteGetUserEntitlementHistory>(AccelByteSubsystem, LocalUserNum, EntitlementId, bForceUpdate);
+}
+
+bool FOnlineEntitlementsAccelByte::GetCachedUserEntitlementHistory(const TSharedRef<const FUniqueNetIdAccelByteUser>& InUserId, const FUniqueEntitlementId& InEntitlementId, TArray<FAccelByteModelsUserEntitlementHistory>& OutEntitlementHistory)
+{
+	FScopeLock ScopeLock(&EntitlementMapLock);
+
+	if (!InUserId->IsValid())
+	{
+		UE_LOG_AB(Warning, TEXT("Failed to get entitlement history from cache as user is invalid"));
+
+		return false;
+	}
+
+	FEntitlementHistory& TempUserEntitlementHistoriesMap = *UserEntitlementHistoryMap.Find(InUserId);
+
+	if (TempUserEntitlementHistoriesMap.Num() == 0)
+	{
+		UE_LOG_AB(Warning, TEXT("Failed to get entitlement history from cache as there are no entitlement histories stored for this user"));
+
+		return false;
+	}
+
+	TArray<FAccelByteModelsUserEntitlementHistory>& UserEntitlementHistories = *TempUserEntitlementHistoriesMap.Find(InEntitlementId);
+
+	if (UserEntitlementHistories.Num() == 0)
+	{
+		UE_LOG_AB(Warning, TEXT("Failed to get entitlement history from cache as there are no history for this entitlement"));
+
+		return false;
+	}
+
+	OutEntitlementHistory = UserEntitlementHistories;
+
+	return true;
+}
+
+bool FOnlineEntitlementsAccelByte::DeleteCachedUserEntitlementHistory(const TSharedRef<const FUniqueNetIdAccelByteUser>& InUserId, const FUniqueEntitlementId& InEntitlementId)
+{
+	FScopeLock ScopeLock(&EntitlementMapLock);
+
+	if (!InUserId->IsValid())
+	{
+		UE_LOG_AB(Warning, TEXT("Failed to delete entitlement history from cache as user is invalid"));
+
+		return false;
+	}
+
+	FEntitlementHistory& TempUserEntitlementHistoriesMap = *UserEntitlementHistoryMap.Find(InUserId);
+
+	if (TempUserEntitlementHistoriesMap.Num() == 0)
+	{
+		UE_LOG_AB(Warning, TEXT("Failed to delete entitlement history from cache as there are no entitlement histories stored for this user"));
+
+		return false;
+	}
+
+	TempUserEntitlementHistoriesMap.Remove(InEntitlementId);
+
+	return true;
 }